requirements.txt.txt
torch==2.4.1
torchvision==0.19.1
opencv-python==4.8.0
pillow==10.0.0
scikit-learn==1.3.0
tqdm==4.66.1